# Faster AVX2 prompt processing for k-quants and IQ4_XS