As discussed [elsewhere](https://github.com/ggerganov/llama.cpp/pull/6840#issuecomment-2079823076), here is a PR that improves AVX2 prompt processing for k-quants and `IQ4_XS` by a large margin. I did not manage to get the speed gains via tinyBlas, so I just added a call in `llamafile_sgemm()` to a separate function that performs the matrix multiplication. 

The table shows a comparison between prompt processing speed on master and with this PR. Not having the `llama-bench` tool here and not knowing how to better measure performance, I just used the `perplexity` tool to measure time for a batch of 512 tokens to get these values. Tested on a 16-core **Ryzen-7950X** CPU with a 7B LLaMA model

| Quants | PP-512 (PR) | PP-512 (master) | Speedup |
|---:|---:|---:|---:|
|Q2_K_S | 157.5 | 111.5 | 1.412 |
|Q3_K_S | 169.8 | 81.3 | 2.089 |
|Q4_K_S | 161.0 | 105.1 | 1.531 |
|Q5_K_S | 146.7 | 72.7 | 2.017 |
|Q6_K | 171.8 | 74.4 | 2.308 |
|IQ4_XS | 147.6 | 74.1 | 1.992 |

For reference, here is what I measure on my system for `fp16` and quants not affected by this PR:

| Quants | PP-512 (master) |
|---:|---:|
| fp16 | 136.5 |
|Q8_0 | 129.6 |
|Q4_0 | 101.4 |
|Q4_1 | 66.8 |
|Q5_0 | 56.5 |
|Q5_1 | 54.2 |

I.e., all k-quants and `IQ4_XS` are now faster than `fp16`!

The speedup in this PR is in most cases better compared to what I reported [here](https://github.com/ggerganov/llama.cpp/pull/6840#issuecomment-2079502775) due to some additional refinements that I have added since this post, but a few percent slower compared to what I get in my private `llama.cpp` fork (with `Q2_K_S` having the most noticeable difference as I get 178 t/s there). Being new to `llamafile`, I'm not sure what is causing such performance differences for the exact same matrix multiplication implementation. 

The same approach as here results in huge performance gains for the other i-quants (`IQ2_XXS, IQ2_XS, IQ2_S, IQ3_XXS, IQ3_S`). But having modified these quants in my repository in ways that make them incompatible with mainline `llama.cpp` i-quants, I have left this part for a future PR.

The Ryzen-7950X implements various parts of the `AVX512` specification. To make sure that this PR provides speedup on non-`AVX512` CPUs, I also tested on an older **Ryzen-5975WX** 32-core CPU. Here I get the following performance for `fp16` and unaffected quants: 

| Quants | PP-512 (master) |
|---:|---:|
| fp16 | 108.7 |
|Q8_0 | 135.1 |
|Q4_0 | 118.8 |

For k-quants and `IQ4_XS` we have

| Quants | PP-512 (PR) | PP-512 (master) | Speedup |
|---:|---:|---:|---:|
|Q2_K_S | 199.2 | 153.7 | 1.296 |
|Q3_K_S | 191.8 | 115.3 | 1.663 |
|Q4_K_S | 182.2 | 140.6 | 1.296 |
|Q5_K_S | 171.8 | 102.8 | 1.671 |
|Q6_K | 187.5 | 107.8 | 1.739 |
|IQ4_XS | 191.8 | 110.6 | 1.734 |