@jart 2024-05-07
This is a remarkable change @ikawrakow. I'm very happy to see that the best quantized formats will now go the fastest. For prompt processing, I'm consistently seeing speedups between 1.2x - 2.0x on x86-64 machines. You even managed to make token generation go faster (which I've found much more difficult), in some cases by as much as 1.33x! Here are my measurements, on three different computers, for three different models.

## Iwan Kawrakow's new GEMM function for K-quants

**Before**: 89c189e9f8212c45621254bce0599e4b49568a4d
**After**: ddb9a8c55281c029961cb0d06a5b43676cbb6ac8

### prompt evalution speed (a.k.a. prefill) in tokens per second

| MODEL          | quant    | microprocessor  | before | after | speedup |
| ---            | ---      | ---             | ---    | ---   | ---     |
| TinyLLaMA 1.1B | Q2\_K    | Intel i9-9900   | 204    | 340   | 1.66x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Intel i9-9900   | 160    | 317   | 1.98x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Intel i9-9900   | 174    | 309   | 1.77x   |
| TinyLLaMA 1.1B | Q4\_0    | Intel i9-9900   | 167    | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Intel i9-9900   | 147    | 280   | 1.90x   |
| TinyLLaMA 1.1B | Q8\_0    | Intel i9-9900   | 219    | -     | -       |
| TinyLLaMA 1.1B | F16      | Intel i9-9900   | 251    | -     | -       |
| TinyLLaMA 1.1B | BF16     | Intel i9-9900   | 222    | -     | -       |
| TinyLLaMA 1.1B | Q2\_K    | Intel i9-14900K | 300    | 600   | 2.00x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Intel i9-14900K | 289    | 606   | 2.10x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Intel i9-14900K | 316    | 606   | 1.92x   |
| TinyLLaMA 1.1B | Q4\_0    | Intel i9-14900K | 418    | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Intel i9-14900K | 275    | 570   | 2.07x   |
| TinyLLaMA 1.1B | Q8\_0    | Intel i9-14900K | 467    | -     | -       |
| TinyLLaMA 1.1B | F16      | Intel i9-14900K | 405    | -     | -       |
| TinyLLaMA 1.1B | BF16     | Intel i9-14900K | 97     | -     | -       |
| TinyLLaMA 1.1B | Q2\_K    | Ryzen 7995WX    | 1350   | 1667  | 1.23x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Ryzen 7995WX    | 1181   | 1648  | 1.39x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Ryzen 7995WX    | 1248   | 1636  | 1.31x   |
| TinyLLaMA 1.1B | Q4\_0    | Ryzen 7995WX    | 1379   | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Ryzen 7995WX    | 961    | 1626  | 1.69x   |
| TinyLLaMA 1.1B | F16      | Ryzen 7995WX    | 1230   | -     | -       |
| TinyLLaMA 1.1B | BF16     | Ryzen 7995WX    | 1800   | -     | -       |
| LLaMA 3 8B     | Q4\_0    | Intel i9-9900   | 27     | -     | -       |
| LLaMA 3 8B     | Q4\_K\_M | Intel i9-9900   | 28     | 41    | 1.46x   |
| LLaMA 3 8B     | Q4\_0    | Intel i9-14900K | 62     | -     | -       |
| LLaMA 3 8B     | Q4\_K\_M | Intel i9-14900K | 57     | 90    | 1.57x   |
| LLaMA 3 8B     | F16      | Intel i9-14900K | 59     | -     | -       |
| LLaMA 3 8B     | Q3\_K\_S | Ryzen 7995WX    | 225    | 416   | 1.84x   |
| LLaMA 3 8B     | Q4\_0    | Ryzen 7995WX    | 278    | -     | -       |
| LLaMA 3 8B     | Q4\_K\_S | Ryzen 7995WX    | 188    | 386   | 2.05x   |
| LLaMA 3 8B     | F16      | Ryzen 7995WX    | 357    | -     | -       |
| LLaMA 3 8B     | BF16     | Ryzen 7995WX    | 508    | -     | -       |
| LLaMA 3 70B    | Q2\_K    | Ryzen 7995WX    | 31     | 51    | 1.65x   |
| LLaMA 3 70B    | Q3\_K\_S | Ryzen 7995WX    | 23     | 44    | 1.91x   |
| LLaMA 3 70B    | Q4\_0    | Ryzen 7995WX    | 31     | -     | -       |
| LLaMA 3 70B    | F16      | Ryzen 7995WX    | 42     | -     | -       |
| LLaMA 3 70B    | BF16     | Ryzen 7995WX    | 65     | -     | -       |

### text generation speed (a.k.a. prediction) in tokens per second

| MODEL          | quant    | microprocessor  | before | after | speedup |
| ---            | ---      | ---             | ---    | ---   | ---     |
| TinyLLaMA 1.1B | Q2\_K    | Intel i9-9900   | 48     | 57    | 1.18x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Intel i9-9900   | 44     | 50    | 1.13x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Intel i9-9900   | 42     | 47    | 1.11x   |
| TinyLLaMA 1.1B | Q4\_0    | Intel i9-9900   | 34     | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Intel i9-9900   | 32     | 35    | 1.09x   |
| TinyLLaMA 1.1B | Q8\_0    | Intel i9-9900   | 25     | -     | -       |
| TinyLLaMA 1.1B | F16      | Intel i9-9900   | 15     | -     | -       |
| TinyLLaMA 1.1B | BF16     | Intel i9-9000   | 15     | -     | -       |
| TinyLLaMA 1.1B | Q2\_K    | Intel i9-14900K | 102    | 129   | 1.26x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Intel i9-14900K | 99     | 125   | 1.26x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Intel i9-14900K | 96     | 113   | 1.17x   |
| TinyLLaMA 1.1B | Q4\_0    | Intel i9-14900K | 86     | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Intel i9-14900K | 74     | 83    | 1.12x   |
| TinyLLaMA 1.1B | Q8\_0    | Intel i9-14900K | 64     | -     | -       |
| TinyLLaMA 1.1B | F16      | Intel i9-14900K | 41     | -     | -       |
| TinyLLaMA 1.1B | BF16     | Intel i9-14900K | 68     | -     | -       |
| TinyLLaMA 1.1B | Q2\_K    | Ryzen 7995WX    | 129    | 160   | 1.24x   |
| TinyLLaMA 1.1B | Q3\_K\_S | Ryzen 7995WX    | 123    | 158   | 1.28x   |
| TinyLLaMA 1.1B | Q3\_K\_M | Ryzen 7995WX    | 122    | 160   | 1.31x   |
| TinyLLaMA 1.1B | Q4\_0    | Ryzen 7995WX    | 129    | -     | -       |
| TinyLLaMA 1.1B | Q5\_K\_M | Ryzen 7995WX    | 109    | 147   | 1.34x   |
| TinyLLaMA 1.1B | F16      | Ryzen 7995WX    | 88     | -     | -       |
| TinyLLaMA 1.1B | BF16     | Ryzen 7995WX    | 79     | -     | -       |
| LLaMA 3 8B     | Q4\_0    | Intel i9-9900   | 6      | -     | -       |
| LLaMA 3 8B     | Q4\_K\_M | Intel i9-9900   | 6      | 6     | 1.00x   |
| LLaMA 3 8B     | Q4\_0    | Intel i9-14900K | 16     | -     | -       |
| LLaMA 3 8B     | Q4\_K\_M | Intel i9-14900K | 15     | 16    | 1.06x   |
| LLaMA 3 8B     | F16      | Intel i9-14900K | 6      | -     | -       |
| LLaMA 3 8B     | Q3\_K\_S | Ryzen 7995WX    | 34     | 46    | 1.35x   |
| LLaMA 3 8B     | Q4\_0    | Ryzen 7995WX    | 37     | -     | -       |
| LLaMA 3 8B     | Q4\_K\_S | Ryzen 7995WX    | 32     | 42    | 1.31x   |
| LLaMA 3 8B     | F16      | Ryzen 7995WX    | 19     | -     | -       |
| LLaMA 3 8B     | BF16     | Ryzen 7995WX    | 20     | -     | -       |
| LLaMA 3 70B    | Q2\_K    | Ryzen 7995WX    | 6      | 8     | 1.33x   |
| LLaMA 3 70B    | Q3\_K\_S | Ryzen 7995WX    | 6      | 7     | 1.16x   |
| LLaMA 3 70B    | Q4\_0    | Ryzen 7995WX    | 5      | -     | -       |
| LLaMA 3 70B    | F16      | Ryzen 7995WX    | 2      | -     | -       |
| LLaMA 3 70B    | BF16     | Ryzen 7995WX    | 2      | -     | -       |