@ikawrakow 2024-05-07
> Looks good to me. Once I get a release out, how would you like to announce it the world? I would like to write a blog post. If you write your own, then I'm happy to tweet that.

I'm not much into blogging, so if you like writing about this, please go ahead.